# w10/server/openapi.yaml
openapi: 3.0.3
info:
  title: EventHub API
  description: REST API for EventHub (events, ticketing, payments, refunds, users, sales)
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegister'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate a user and return JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLogin'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout current user (invalidate token)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Logged out
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /events/{event_id}/ticket-types:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [TicketTypes]
      summary: List ticket types for an event
      responses:
        '200':
          description: Ticket types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketTypeResponse'
    post:
      tags: [TicketTypes]
      summary: Create ticket type for event
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketTypeCreate'
      responses:
        '201':
          description: Ticket type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketTypeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /payments/{payment_id}/refunds:
    parameters:
      - $ref: '#/components/parameters/PaymentId'
    post:
      tags: [Refunds]
      summary: Initiate refund for a payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundCreate'
      responses:
        '201':
          description: Refund requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /refunds/{refund_id}:
    parameters:
      - in: path
        name: refund_id
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Refunds]
      summary: Get refund status
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Refund info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Refunds]
      summary: Update refund (admin processing)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: processed
      responses:
        '200':
          description: Refund updated
  /tickets:
    get:
      tags: [Tickets]
      summary: List tickets (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: event_id
          schema:
            type: string
            format: uuid
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Tickets list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponse'
  /users/{user_id}/tickets:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Tickets]
      summary: List tickets for a user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TicketResponse'
  /reservations/{reservation_id}/tickets:
    parameters:
      - $ref: '#/components/parameters/ReservationId'
    post:
      tags: [Tickets]
      summary: Issue ticket for reservation (after payment)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Ticket issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    EventId:
      name: event_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ReservationId:
      name: reservation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PaymentId:
      name: payment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: Validation failed
    AuthRegister:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: newuser@example.com
        password:
          type: string
          example: strongP@ssw0rd
        full_name:
          type: string
          example: New User
    AuthLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: bob@example.com
        password:
          type: string
          example: hash2
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/UserResponse'
...
    UserResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        email:
          type: string
          example: alice@example.com
        full_name:
          type: string
          example: Alice Smith
        status:
          type: string
          example: active
        created_at:
          type: string
          format: date-time
          example: 2026-03-01T12:00:00Z
...
    TicketTypeCreate:
      type: object
      required: [name, price, quantity]
      properties:
        name:
          type: string
          example: VIP
        price:
          type: number
          format: decimal
          example: 15000
        quantity:
          type: integer
          example: 10
        metadata:
          type: object
...
    TicketTypeResponse:
      type: object
      properties:
        ticket_type_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        metadata:
          type: object
...
    TicketResponse:
      type: object
      properties:
        ticket_id:
          type: string
          format: uuid
        ticket_type_id:
          type: string
          format: uuid
        reservation_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        seat:
          type: string
        status:
          type: string
        issued_at:
          type: string
          format: date-time
...
    RefundCreate:
      type: object
      required: [amount, reservation_id]
      properties:
        reservation_id:
          type: string
          format: uuid
        amount:
          type: number
        reason:
          type: string
    RefundResponse:
      type: object
      properties:
        refund_id:
          type: string
          format: uuid
        payment_id:
          type: string
          format: uuid
        reservation_id:
          type: string
          format: uuid
        amount:
          type: number
        status:
          type: string
        requested_at:
          type: string
          format: date-time
...